# Improve `parse_utf8` performance