Improve parseUtf8 performance.

All the timings are using the rdtsc instruction, the lower the better.

 Timings using a 9mb GLTF file from this [issue](https://github.com/godotengine/godot/issues/98893) averaged on 6000 calls of the function. **Old len** is the current implementation with the length of the string passed in, **Old -1**  is the current implementation without the length of the string passed in. **New** is the proposed implementation.
![GLTF_FileComparison](https://github.com/user-attachments/assets/a0226770-03f1-452a-8c93-2fb56f244e9c)
Went from 3 083 895 cycles down to 1 816 624 cycles.

Timings using a 6mb Ascii file from [here](https://github.com/dscape/spell/blob/master/test/resources/big.txt) averaged on 15 000 calls of the function.
![UTF8File_comparison](https://github.com/user-attachments/assets/83c24d50-56f4-4540-a6bd-47a7b88a80dd)
Went from 50 741 668 cycles down to 30 046 082 cycles.

Tested the output using [utf8tests](https://github.com/flenniken/utf8tests/tree/main). Now passes all the section, except the _Null Characters_ section (being able to have a 0 byte inside a longer string), but this is to be expected as changing this would probably break a lot of usage code.

Regarding tests, I had to remove the non standard ones, as it is now closer to standard. I replaced them with examples from the [unicode standard specification](https://www.unicode.org/versions/Unicode16.0.0/UnicodeStandard-16.0.pdf). 

Some tests from other people would be nice, as this seems pretty widely used through the engine.