@Ivorforce 2024-11-29
Regarding `NULL` termination, I'll link my proposal from just yesterday: https://github.com/godotengine/godot-proposals/issues/11249

Agree that's for another day. Thanks for thoroughly testing this PR! Can you explain the changes on a high level?