@kiroxas 2024-11-29
> Regarding `NULL` termination, I'll link my proposal from just yesterday: [godotengine/godot-proposals#11249](https://github.com/godotengine/godot-proposals/issues/11249)
> 
> Agree that's for another day. Thanks for thoroughly testing this PR! Can you explain the changes on a high level?

Nothing fancy, just do everything in one pass instead of two, and some small look ahead instead of one char per iteration with storing some state.