@clayjohn 2024-11-29
> Regarding tests, I had to remove the non standard ones, as it is now closer to standard. I replaced them with examples from the [unicode standard specification](https://www.unicode.org/versions/Unicode16.0.0/UnicodeStandard-16.0.pdf).

Does this mean that the new implementation fails the previous tests?