@bruvzg 2024-11-29
> Does this mean that the new implementation fails the previous tests?

Previous implementation was decoding overlongs, now one do not (can be changed by removing few `unicode = _replacement_char;` lines).

We also might want to decode unpaired surrogates, previous implementation was not doing it since #74760, but unpaired surrogates can be par of Windows file names, and we probably should be able to store/read them (not sure how common it is).

Both cases should still set the parsing error flag (since it's not valid UTF-8).