@kiroxas 2024-11-29
> > Does this mean that the new implementation fails the previous tests?
> 
> Previous implementation was decoding overlongs, new one do not (can be changed by removing few `unicode = _replacement_char;` lines).
> 
> We also might want to decode unpaired surrogates, previous implementation was not doing it since #74760, but unpaired surrogates can be par of Windows file names, and we probably should be able to store/read them (not sure how common it is).
> 
> Both cases should still set the parsing error flag (since it's not valid UTF-8).



> > Does this mean that the new implementation fails the previous tests?
> 
> Previous implementation was decoding overlongs, new one do not (can be changed by removing few `unicode = _replacement_char;` lines).
> 
> We also might want to decode unpaired surrogates, previous implementation was not doing it since #74760, but unpaired surrogates can be par of Windows file names, and we probably should be able to store/read them (not sure how common it is).
> 
> Both cases should still set the parsing error flag (since it's not valid UTF-8).

Seems like a user can manually name a file with an unpaired surrogate and windows will accept it. I see other issues in programming languages like [zig](https://github.com/ziglang/zig/issues/2565) and [go](https://github.com/golang/go/issues/32334) mitigating this by using [WTF-8 ](https://simonsapin.github.io/wtf-8/) instead of UTF-8. This would be a small change, it consists mostly of [accepting surrogates](https://simonsapin.github.io/wtf-8/#encoding-wtf-8) with a small transformation. This is probably better suited for another PR after this one though.